backend/pyproject.toml
[project]
name = "roma01"
version = "1.1.0"
description = "ROMA-01: AI-powered cryptocurrency futures trading platform"
readme = "README.md"
requires-python = ">=3.12,<3.14"
license = {text = "MIT"}
authors = [
    {name = "Luke Ma", email = "lukema95@github.com"}
]
keywords = ["ai", "trading", "crypto", "futures", "dspy", "llm", "roma"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "dspy>=3.0.3",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.11.0",
    "sqlalchemy>=2.0.35",
    "aiosqlite>=0.20.0",
    "httpx>=0.26.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.3",
    "pyyaml>=6.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "ta-lib>=0.4.28",
    "websockets>=15.0.1",
    "eth-account>=0.13.0",
    "eth-abi>=5.1.0",
    "web3>=7.0.0",
]

[project.urls]
Homepage = "https://github.com/lukema95/roma-01"
Documentation = "https://github.com/lukema95/roma-01/tree/main/docs"
Repository = "https://github.com/lukema95/roma-01"
Issues = "https://github.com/lukema95/roma-01/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/roma_trading"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

